// src/gen/openapi.ts
export interface components {
    schemas: {
// ...
        MarketsInfo: {
            markets: components["schemas"]["MarketInfo"][];
            tokens: components["schemas"]["TokenInfo"][];
        };
// ...
        ActionsItem: {
            /** Format: uint64 */
            actionId: number;
            physicalTime: string;
            payload: string;
        };
// ...
        FillRole: "maker" | "taker";
        UserNotFound: null;
        OrderbookInfo: {
            /** Format: uint64 */
            updateId: number;
            asks: [
                number,
                number
            ][];
            bids: [
                number,
                number
            ][];
            asksSummary: components["schemas"]["SideSummary"];
            bidsSummary: components["schemas"]["SideSummary"];
        };
        SideSummary: {
            /** Format: double */
            sum: number;
            /** Format: uint32 */
            count: number;
        };
        MarketNotFound: {
            /** Format: uint32 */
            market_id: number;
        };
        MarketStats: {
            /** Format: double */
            indexPrice: number;
            /** Format: double */
            indexPriceConf: number;
            frozen?: boolean;
            /** Format: double */
            volumeBase24h: number;
            /** Format: double */
            volumeQuote24h: number;
            /** Format: double */
            high24h: number;
            /** Format: double */
            low24h: number;
            /** Format: double */
            close24h: number;
            /** Format: double */
            prevClose24h?: number | null;
            perpStats?: components["schemas"]["PerpMarketStats"] | null;
        };
        PerpMarketStats: {
            /** Format: double */
            mark_price?: number | null;
            /** Format: double */
            aggregated_funding_index: number;
            /** Format: double */
            funding_rate: number;
            next_funding_time: string;
            /** Format: double */
            open_interest: number;
        };
        TokenStats: {
            symbol: string;
            /** Format: uint8 */
            decimals: number;
            mintAddr: string;
            /** Format: uint16 */
            weightBps: number;
            oracleSymbol: string;
            /** @description Index price for the token */
            indexPrice?: components["schemas"]["TokenPrice"] | null;
        };
        TokenPrice: {
            /** Format: double */
            median: number;
            /** Format: double */
            confidence: number;
        };
        TokenNotFound: null;
        User: {
            accountIds: number[];
            sessions: {
                [key: string]: components["schemas"]["UserSession"];
            };
        };
        UserSession: {
            pubkey: string;
            expiry: string;
        };
// ...
        PageResult_for_String_and_OrderInfo: {
            /** @description Set of items for requested by query. */
            items: components["schemas"]["OrderInfo"][];
            /** @description If request contains more data, this is the id is set with which next request should be performed to get next page. If no more data, then it is undefined. */
            nextStartInclusive?: string | null;
        };
        /** @description Order summary. Use order_id to application side join with trades as needed. */
        OrderInfo: {
            addedAt: string;
            updatedAt: string;
            tradeId: string;
            /** Format: uint32 */
            traderId: number;
            /** Format: uint32 */
            marketId: number;
            orderId: string;
            side: components["schemas"]["Side"];
            /** Format: double */
            placedSize: number;
            /** Format: double */
            filledSize?: number | null;
            updateActionId: string;
            isReduceOnly: boolean;
            fillMode: components["schemas"]["FillMode"];
            /** Format: double */
            placedPrice: number;
            /** Format: double */
            originalSizeLimit?: number | null;
            /** Format: double */
            originalPriceLimit?: number | null;
            placementOrigin: components["schemas"]["PlacementOrigin"];
            finalizationReason?: components["schemas"]["FinalizationReason"] | null;
            marketSymbol: string;
            tokenSymbol: string;
        };
        Side: "ask" | "bid";
        FillMode: "Limit" | "PostOnly" | "ImmediateOrCancel" | "FillOrKill";
        /** @enum {string} */
        PlacementOrigin: "User" | "Trigger" | "Liquidation";
        /** @enum {string} */
        FinalizationReason: "Filled" | "Canceled" | "Taken";
        Account: {
            /** Format: uint64 */
            updateId: number;
            orders: components["schemas"]["OpenOrder"][];
            positions: components["schemas"]["PositionSummary"][];
            balances: components["schemas"]["Balance"][];
            margins: components["schemas"]["AccountMarginsView"];
        };
        OpenOrder: {
            orderId: string;
            /** Format: uint32 */
            marketId: number;
            side: components["schemas"]["Side"];
            /** Format: double */
            size: number;
            /** Format: double */
            price: number;
            /** Format: double */
            originalOrderSize: number;
            /** Format: uint64 */
            clientOrderId?: number | null;
        };
        /** @description Summary of position. Does not provide details information about specific orders. */
        PositionSummary: {
            /** Format: uint32 */
            marketId: number;
            /** Format: uint16 */
            openOrders: number;
            perp?: components["schemas"]["PerpPosition"] | null;
            /** Format: uint64 */
            actionId: number;
        };
        PerpPosition: {
            /** Format: double */
            baseSize: number;
            /** Format: double */
            price: number;
            /** Format: double */
            updatedFundingRateIndex: number;
            /**
             * Format: double
             * @description In stable coin, USDC
             */
            fundingPaymentPnl: number;
            /**
             * Format: double
             * @description In stable coin, USDC
             */
            sizePricePnl: number;
            isLong: boolean;
        };
        Balance: {
            /** Format: uint32 */
            tokenId: number;
            token: string;
            /** Format: double */
            amount: number;
        };
        /** @description Represents account health state. Each field expressed in USD basis points per account. For example: - value of 1.0000 means 1 USD. - value of 123.0004 means 123 USD and 0.04 cent.
         *
         *     All values are actual only at specific moment in time - because both market and oracle prices fluctuate all the time */
        AccountMarginsView: {
            /**
             * Format: double
             * @description Somewhat proportional to USD weighted value of account, its tokens and positive PnL; sub negative PnL, debt.
             *
             *     See MARKETS.md for details. Specifically have MF as basis points, divide by `pn`
             */
            omf: number;
            /**
             * Format: double
             * @description Similar to OMF. See distinguishing in MARKETS.md.
             *
             *     Specifically have MF as basis points, divide by `pon`.
             */
            mf: number;
            /**
             * Format: double
             * @description Compare with OMF. Divide by PON to get basis points value. Tells when account may or may not open orders.
             */
            imf: number;
            /**
             * Format: double
             * @description Compare with OMF. Divide by PON to get basis points value. Tells when account's orders to be cancelled.
             */
            cmf: number;
            /**
             * Format: double
             * @description Compare with MF. Divide by PON to get basis points value. Tells when account positions reduced and assets sold.
             */
            mmf: number;
            /**
             * Format: double
             * @description Expressed in USD, can be zero. See PON in MARKETS.md for details. Somewhat proportional to open position size and orders size in OB.
             */
            pon: number;
            /**
             * Format: double
             * @description Similar to `pon`. See PN in MARKETS.md.
             */
            pn: number;
            /** @description See `bankruptcy` in MARKETS.md. In general happens if account does not have enough weighted tokens to covert his debt and negative PnL. */
            bankruptcy: boolean;
        };
        PagedQuery: {
            /**
             * @description start with this timestamp (RFC3339); defaults to UNIX epoch start
             * @default null
             */
            since: string;
            /**
             * @description end with this timestamp (RFC3339); defaults to current date-time
             * @default null
             */
            until: string;
            /** @description fetch results starting with this page; query starts with first entry if page isn't specified */
            startInclusive?: string | null;
            /**
             * Format: uint8
             * @description Query returns up to 50 trades in one go.
             * @default null
             */
            pageSize: number | null;
        };
// ...
        PageResult_for_String_and_Trade: {
            /** @description Set of items for requested by query. */
            items: components["schemas"]["Trade"][];
            /** @description If request contains more data, this is the id is set with which next request should be performed to get next page. If no more data, then it is undefined. */
            nextStartInclusive?: string | null;
        };
        Trade: {
            time: string;
            actionId: string;
            tradeId: string;
            /** Format: uint32 */
            takerId: number;
            takerSide: components["schemas"]["Side"];
            /** Format: uint32 */
            makerId: number;
            /** Format: uint32 */
            marketId: number;
            orderId: string;
            /** Format: double */
            price: number;
            /** Format: double */
            baseSize: number;
        };
// ...
        FeeTierId: number;
        FeeTierConfig: {
            /** Format: uint16 */
            maker_fee_bps: number;
            /** Format: uint16 */
            taker_fee_bps: number;
        };
// ...
        PageResult_for_uint32_and_AccountFeeTier: {
            /** @description Set of items for requested by query. */
            items: components["schemas"]["AccountFeeTier"][];
            /**
             * Format: uint32
             * @description If request contains more data, this is the id is set with which next request should be performed to get next page. If no more data, then it is undefined.
             */
            nextStartInclusive?: number | null;
        };
        AccountFeeTier: {
            /** Format: uint32 */
            accountId: number;
            feeTier: components["schemas"]["FeeTierId"];
        };
        AdminInfo: {
            key: string;
            roles: string[];
        };
        GetAccountVolumeQuery: {
            /** Format: uint32 */
            accountId: number;
            /** Format: uint32 */
            marketId?: number | null;
            /**
             * @description start with this timestamp (RFC3339); defaults to UNIX epoch start
             * @default null
             */
            since: string;
            /**
             * @description end with this timestamp (RFC3339); defaults to current date-time
             * @default null
             */
            until: string;
        };
        AccountVolumeInfo: {
            /** Format: uint32 */
            marketId: number;
            /** Format: double */
            volumeBase: number;
            /** Format: double */
            volumeQuote: number;
        };
        GetPrevMarketPriceQuery: {
            atOrBefore: string;
        };
        PreviousMarketPrice: {
            /** Format: uint32 */
            marketId: number;
            /** Format: double */
            price?: number | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}